producer/requirements.txt
requests>=2.31.0
kafka-python>=2.0.2
python-dateutil>=2.8.2
